Renamed files to be more pythonic.
