+++ pyum/rpm.py
@@ -1,7 +1,7 @@
 # import yum
 import tempfile
 from pyrpm.rpm import RPM
-from pyum.HTTPClient import HTTPClient
+from pyum.httpclient import HTTPClient
 
 __author__ = 'drews'
